feat: WIP add WaaS support
