+++ internal/api/common/common.go
@@ -4,3 +4,9 @@ type Secret struct {
 	Encrypted string `json:"encrypted,omitempty"`
 	Plain     string `json:"plain,omitempty"`
 }
+
+type PortRange struct {
+	Deny  bool `json:"deny,omitempty"`
+	End   int  `json:"end,omitempty"`
+	Start int  `json:"start,omitempty"`
+}
